requirements.txt
pytest
pytest-cov
pandas
yagmail
python-dotenv
